# Conflicts:
#   project/app/Http/Controllers/AdminController.php
#   project/public/css/admin.css

+++ project/app/Http/Controllers/AdminController.php
@@ -26,6 +26,8 @@ class AdminController extends Controller
      */
     public function create()
     {
+        $user = auth()->user();
+        $this->authorize('admin', $user);
         return view('admin.create');
     }
 
@@ -34,6 +36,8 @@ class AdminController extends Controller
      */
     public function store(Request $request)
     {
+        $user = auth()->user();
+        $this->authorize('admin', $user);
         $validated = $request->validate([
             'name' => 'required|string|max:255',
             'email' => 'required|email',
@@ -60,6 +64,8 @@ class AdminController extends Controller
      */
     public function edit(string $id)
     {
+        $loguser = auth()->user();
+        $this->authorize('admin', $loguser);
         $user = User::findOrFail($id);
 
         $this->authorize('update', $user);
@@ -80,6 +86,8 @@ class AdminController extends Controller
      */
     public function destroy(string $id)
     {
+        $loguser = auth()->user();
+        $this->authorize('admin', $loguser);
         $user = User::findOrFail($id);
         $user->delete();
 
@@ -132,8 +140,8 @@ class AdminController extends Controller
      */
     public function showUser(string $id)
     {
-        $user = auth()->user();
-        $this->authorize('admin', $user);
+        $loguser = auth()->user();
+        $this->authorize('admin', $loguser);
         $user = User::findOrFail($id);
         return view('admin.showUser', compact('user'));
     }
@@ -143,8 +151,8 @@ class AdminController extends Controller
      */
     public function showGroup(string $id)
     {
-        $user = auth()->user();
-        $this->authorize('admin', $user);
+        $loguser = auth()->user();
+        $this->authorize('admin', $loguser);
         $group = Group::findOrFail($id);
         return view('admin.showGroup', compact('group'));
     }
@@ -154,11 +162,9 @@ class AdminController extends Controller
      * Ban a user
      */
     public function banUser(string $id){
-        $user = auth()->user();
-        $this->authorize('admin', $user);
+        $loguser = auth()->user();
+        $this->authorize('admin', $loguser);
         $user = User::findOrFail($id);
-        
-
         return redirect()->route('admin.users.manage')->with('success', 'User banned successfully.');
     }
 
@@ -167,9 +173,8 @@ class AdminController extends Controller
      */
     public function changePassword()
     {
-        $user = auth()->user();
-        $this->authorize('admin', $user);
-        Log::info('Showing change password page');
+        $loguser = auth()->user();
+        $this->authorize('admin', $loguser);
         return view('admin.changePassword');
     }
 
@@ -178,13 +183,19 @@ class AdminController extends Controller
      */
     public function updatePassword(Request $request, string $id)
     {
-        $user = auth()->user();
-        $this->authorize('admin', $user);
+        $loguser = auth()->user();
+        $this->authorize('admin', $loguser);
         $validatedData = $request->validate([
             'old' => 'required|string|min:8',
             'new' => 'required|string|min:8',
+            'repeat' => 'required|string|min:8',
         ]);
         $user = User::findOrFail($id);
+
+        if ( $validatedData['new'] != $validatedData['repeat'] ) {
+            return redirect()->route('admin.changePassword', $id)->with('error', 'Passwords did not match.');
+        }
+
 
         if (Hash::check($validatedData['old'], $user->password)) {
             $user->password = bcrypt($validatedData['new']);
