+++ project/app/Http/Controllers/UserController.php
@@ -13,6 +13,8 @@ class UserController extends Controller
      */
     public function index()
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $users = User::all();
         return view('users.index', compact('users'));
     }
@@ -22,6 +24,8 @@ class UserController extends Controller
      */
     public function create()
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         return view('users.create');
     }
 
@@ -30,6 +34,8 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $validated = $request->validate([
             'name' => 'required|string|max:255',
             'email' => 'required|email',
@@ -48,6 +54,8 @@ class UserController extends Controller
      */
     public function show(string $id)
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $user = User::findOrFail($id);
 
         if (!$user->is_public && !auth()->check()) {
@@ -75,6 +83,8 @@ class UserController extends Controller
 
     public function follow(Request $request)
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $userToFollow = User::findOrFail($request->user2_id);
         $loggedInUser = auth()->user();
 
@@ -97,43 +107,65 @@ class UserController extends Controller
 
     public function checkRequests()
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $user = auth()->user();
         $pendingRequests = \App\Models\Follow::with('follower')
                                             ->where('user2_id', $user->id)
                                             ->where('request_status', 'pending')
                                             ->get();
 
-        return view('requests.show', compact('pendingRequests','user'));
+        $pendingRequestsCount = $pendingRequests->count();
+        return view('requests.show', compact('pendingRequests','pendingRequestsCount','user'));
     }
 
     public function accept($user1_id, $user2_id)
     {
-        // Find the follow request using both user1_id and user2_id
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $request = \App\Models\Follow::where('user1_id', $user1_id)
                                     ->where('user2_id', $user2_id)
-                                    ->first();
+                                    ->delete();
 
         if ($request) {
-            $request->request_status = 'accepted';
-            $request->save();
+            \App\Models\Follow::create([
+                'user1_id' => $user1_id,
+                'user2_id' => $user2_id,
+                'request_status' => 'accepted',
+            ]);
 
-            return redirect()->route('requests.show')->with('success', 'Follow request accepted.');
+            return redirect()->route('home')->with('success', 'Request accepted.');
         }
 
-        return redirect()->route('requests.show')->with('error', 'Follow request not found.');
+        return redirect()->route('home')->with('error', 'Request not found.');
     }
 
     public function reject($user1_id, $user2_id)
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $deleted = \App\Models\Follow::where('user1_id', $user1_id)
                                     ->where('user2_id', $user2_id)
                                     ->delete();
     
         if ($deleted) {
-            return redirect()->route('requests.show')->with('success', 'Follow request rejected.');
+            return redirect()->route('home')->with('success', 'Request rejected.');
         }
     
-        return redirect()->route('requests.show')->with('error', 'Follow request not found.');
+        return redirect()->route('home')->with('error', 'Request not found.');
+    }
+
+    public function unfollow(Request $request)
+    {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
+
+        \App\Models\Follow::where([
+            ['user1_id', '=', $request->user1_id],
+            ['user2_id', '=', $request->user2_id],
+        ])->delete();
+
+        return response()->json(['message' => 'Unfollowed successfully'], 200);
     }
 
     /**
@@ -141,6 +173,8 @@ class UserController extends Controller
      */
     public function edit(string $id)
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $user = User::findOrFail($id);
 
         $this->authorize('update', $user);
@@ -152,6 +186,8 @@ class UserController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         // Validate the incoming data
         $validatedData = $request->validate([
             'name' => 'required|string|max:255',
@@ -198,6 +234,8 @@ class UserController extends Controller
      */
     public function destroy(string $id)
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $user = User::findOrFail($id);
         $user->delete();
 
@@ -205,34 +243,12 @@ class UserController extends Controller
     }
 
     /**
-     * Delete a user making it a anonymous user
-     */
-    public function deleteUser(string $id)
-    {
-        $user = User::findOrFail($id);
-        $user->username = null;
-        $user->firstname = null;
-        $user->surname = null;
-        $user->password = null;
-        $user->email = null;
-        $user->bio_description = null;
-        $user->profile_picture = 'default.jpg';
-        $user->is_public = false;
-        $user->type = 'deleted';
-        $user->save();
-
-        return redirect()->route('admin.users.manage')->with('success', 'User deleted successfully.');
-
-    }
-
-    
-
-
-    /**
      * Show the settings page.
      */
     public function settings()
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $user = auth()->user();
         return view('settings.show', compact('user'));
     }
@@ -242,6 +258,8 @@ class UserController extends Controller
      */
     public function updatePassword(Request $request, string $id)
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $validatedData = $request->validate([
             'old' => 'required|string|min:8',
             'new' => 'required|string|min:8',
@@ -263,6 +281,10 @@ class UserController extends Controller
      */
     public function deleteUser(string $id)
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $user = User::findOrFail($id);
         $user->username = null;
         $user->firstname = null;
@@ -287,6 +309,8 @@ class UserController extends Controller
      */
     public function privacity(Request $request, string $id)   
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $user = User::findOrFail($id);
 
         $validatedData = $request->validate([
